refactor(plugin): move payload types to shared package and sanitize error messages
- Move PluginInvokePayload, PluginSetEnabledPayload, ProviderInfo to shared/payloads.ts
- Add CliDetectionStatus transport type in shared (structurally compatible with CliDetectionResult)
- Remove duplicate ProviderInfo from frontend plugin store
- Use generic error messages for pluginId/activation errors (no user-input reflection)
